Authoraization Completion

+++ app/Http/Controllers/PostsController.php
@@ -7,11 +7,21 @@ use App\Post;
 use DB;
 class PostsController extends Controller
 {
+        /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth',['except'=> ['index','show']]);
+    }
     /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
      */
+    
     public function index()
     {
 //$posts=DB::select('SELECT * FROM posts');
@@ -82,6 +92,10 @@ class PostsController extends Controller
     {
         //
         $post =  Post::find($id);
+         //Checking for correct user
+        if(auth()->user()->id!==$post->user_id){
+            return redirect('/posts')->with('error','Unauthorized Page!!!!!!!!');
+        }
         return view('posts.edit')->with('post',$post);
     
     }
@@ -120,6 +134,10 @@ class PostsController extends Controller
     {
         //
         $post = Post::find($id);
+        //Checking for correct user
+        if(auth()->user()->id!==$post->user_id){
+            return redirect('/posts')->with('error','Unauthorized Page!!!!!!!!');
+        }
         $post->delete();
         return redirect('/posts')->with('success','Post Removed');
